New Branch / New Folder Layout
This commit is to push a new folder layout of how I would want the
structure of the program to be. It is intended to be a set up of what I
want for the feature of the coding side of the project to be. Nothing
will compile since I have not changed the code to incorporate the new
design. I shall work on this in the future.