Added functionality to the Fly and Swarm Classes.
I have cleaned the getContourFrame function to only return the contour
made by evaluateContours. This is more efficient so that even when the
program isn't displaying contours it is still tracking in the
background. We shouldn't need to have to click 'c' on the keyboard to
start tracking anymore.

Next is now to persistantly track each fly, by comparing one frame to
the last to find the closest fly.