Added a Max Flies Limiter...
The program now has a set limit of flies that are in a vile.

The default is set to 12, but will soon be able to be changed through a
slider / textbox.

Have cleaned the code since there were duplicate functions that had done
the same thing; (just under different names).