Performance upgrade.
Main loop now always gets a normal frame to show. Then if contours are detected, the contours will be shown on top of the normal frame. Also I have made a new function called check_windows that will detect if the 'q' key has been pressed to exit the program.