Huge changes. Class overhaul and multithreading.
I have completely changed the videostream class into what is now the
FlyContourTracker class. I have optomized the tracking to track at a
frame rate of 15 fps. The higher the resolution the slower the program
gets, since the raspberypi is just a small ARM computer. The default
will be (1280, 720) resolution. I have better commented the code,
multi-threaded the program, and given better output feedback for
debugging. The code will soon support multiple viles again, and will
soon be given the ability to 'start' and 'end' sessions for individual
tracking.

Now to code only takes 2 functions to run. One to start the process to
take frames from the camera, and a second to process the contours within
the camera.