Major Changes to structure
Firstly, the program now takes the image from the camera, converts it
into grayscale in another thread, and then 4 other different threads now
take segments of the grayscale image to process. The program now
preforms at a faster rate since each thread has a smaller image to
process. The smaller images should be of size (320, 720) when segmenting
the image into 4 seprate images. The process of finding contours is the
same as the previous update. Also I have added more flexability in the
ability to turn on and off sections of the program for easier debuging.
The motor class now has an boolean identifier to know of the motor is on
or off. This allows my program to avoid collision of two different
threads in the event that two different threads need to shake the vile.
Added the ability to use USB cameras just incase. lastly I have changed
the way the program requests a frame from the camera using OpenCV's
Capture_Sequence function. It will now capture at a rate of 10 fps.